feat: return repo `fullname` in fork script (#29)
This PR is a follow-up to a missing step in the `fork` script initial
implementation at #3; the fork script failed to return a `repo` which
can be used to in the next step of computation. This was because of a
weird assumption I had during the initial implementation. 😆See my
assumption below...

> I assume that the call to the `"POST /repos/{owner}/{repo}/forks"`
endpoint only assures of initiating a fork process without assuring us
of a response at all. Meaning we might not exactly get a `response.data`
following the call

...but that wasn't true, I found out that a `response.data` actually
comes, but it might just take some time and only in cases where the repo
being forked is huge.... and at the moment forking the project repo
happens in less than 5secs.

### Changes Made

- Returned `fork` repo - this is a repo fullname value returned from the
`isRepositoryForked` helper function; I hereby return it as main
returned value from the `forkRepository` function execution in the
condition where the repo is already forked on a executing user's account
- Returned `response.data.full_name` - this is a newly created fork repo
fullname; Its a value from the response to the `"POST
/repos/{owner}/{repo}/forks"` endpoint call; I hereby return it as main
retuned value from the `forkRepository` function execution in cases
where there was no fork already already found on the executing user's
account
- Cherry picked some changes from #25 to use on here
  - f12f25f548a5c5836e9be7d601ed226c5269f5ee
  - 436ceea649b67812c0ec1164fde95d443ce556e0

📖