fix: dictionary search engine late indexing (#49)
This Pull Request fixes the issue where the dictionary search engine
returns no result on initial load of the web app. This is exactly due to
the reason explained in #47. This fix implements direct computation of
the `dictionary` object property (derived from an `Astro.glob` operation
used to retrieve all entries from our words directory `pages/browse`)
inside of the homepage and the word layout, dropping usage of state to
share the value.

### Changes Made 

- Deprecated and dropped the integration of the `$dictionary` state 
- Compute value for the `search` island `dictionary` prop (which used to
be set to the `$dictionary.value`) directly in the server-side of the
following places
- `layouts/word.astro` - the word layout where the `search` island is
integrated in the `Navbar`
- `pages/index.astro` - the landing page where the `search` island is
prominently integrated

### Related Issue

Resolves #47 

### Screencast/Screenshot

[screencast-jargons-dev-git-fix-dictionary-search-index-babblebeys-projects.vercel.app-2024.04.19-11_26_47.webm](https://github.com/babblebey/jargons.dev/assets/25631971/9c500897-63f3-4203-b604-372ee056527b)

📖