fix(build): Move '@types/react' to devDependencies (#92)
### Description
This PR removes the `@type/react` dependency from `dependencies` and
adds it to the `devDependencies` in the projects `package.json` file


### Related Issue
Fixes #21 

### Screenshots/Screencasts

![image](https://github.com/user-attachments/assets/74f80ca8-2c93-47b2-a612-c2ab00b80268)


### Notes to Reviewer
I initially don't know what happened in the first commit. I ran `npm
install` after moving `@types/react` dependency to the the
devDependency. But I did it well in the second commit.

---------

Co-authored-by: Olabode Lawal-Shittabey <babblebey@gmail.com>