feat: implement jargonsdev logo (#124)
### Description
<!-- Please add PR description (don't leave blank) - example: This PR
[adds/removes/fixes/replaces] the [feature/bug/etc] -->

This Pull Request introduces the `jargons.dev` logotype to the homepage
and in the navigation bar on the word layout and jargons editor.

**Changes Made:**

- Introduced the `JargonsdevLogo` component which returns an svg content
of the logo; it accepts a class prop that drills down to the svg's class
attribute for better styling.
- Integrated the `JargonsdevLogo` component in intended places
- Modified navigation return button to use icon only mode, relegating
the `returnLabel` text to screen-reader only
- Modified the strokeWidth of navigation return button and search form
icon for better visibility

### Related Issue
<!-- Please prefix the issue number with Fixes/Resolves - example: Fixes
#123 or Resolves #123 -->

NA

### Screenshots/Screencasts
<!-- Please provide screenshots or video recording that demos your
changes (especially if it's a visual change) -->


[screencast-localhost_4321-2024_12_06-20_37_38.webm](https://github.com/user-attachments/assets/f4ddb6d3-e4de-440c-8dc2-75f36efa5177)

### Notes to Reviewer
<!-- Please state here if you added a new npm packages, or any extra
information that can help reviewer better review you changes -->

NA