Dictionary (New Word): Infinite Loop (#84)
# Word

#### Infinite Loop

# Meaning/Definition

An infinite loop is a loop that has no end. A loop that has a
terminating condition that can never be met. To some certain conditions
it can be useful but most times causes issues when not properly handled.
Example of an infinite loop in JS is a while loop

while (true) {
    console.log("This will run forever.");
}

---------

Co-authored-by: Olabode Lawal-Shittabey <babblebey@gmail.com>