doc(contribution-guide): add guidelines for issues closed to community pickup (#222)
## Description
<!-- Please add PR description (don't leave blank) - example: This PR
[adds/removes/fixes/replaces] the [feature/bug/etc] -->

This pull request updates the `CONTRIBUTING.md` file to clarify which
issues are not open for community assignment and adds guidelines for
interacting with such issues. The goal is to help contributors
understand when they should not request assignment or submit pull
requests for certain labeled issues.

**Updates to issue assignment guidelines:**

* Added a new section describing the `maintainers-only` and `wontfix`
labels, explaining that issues with these labels are not open for
general contributor assignment and specifying the reasons behind these
restrictions.
* Provided clear guidelines on how to interact with issues labeled as
unavailable, including advice on asking for clarification and suggesting
alternative approaches.

## Related Issue
<!-- Please prefix the issue number with Fixes/Resolves - example: Fixes
#123 or Resolves #123 -->

NA

## Screenshots/Screencasts
<!-- Please provide screenshots or video recording that demos your
changes (especially if it's a visual change) -->

NA

## Notes to Reviewer
<!-- Please state here if you added a new npm packages, or any extra
information that can help reviewer better review you changes -->

NA