fix(profile): improve dropdown button styling and visual cues (#250)
## Description
<!-- Please add PR description (don't leave blank) - example: This PR
[adds/removes/fixes/replaces] the [feature/bug/etc] -->

This pull request updates the `Profile` component to improve the visual
behavior and accessibility of the profile button and its dropdown
indicator. The most notable changes focus on the button's styling and
the dynamic display of open/close icons.

**UI/UX improvements to profile button:**

* The profile button's ring color and background have been adjusted for
better contrast and consistency, now using a conditional ring color and
a lighter background. (`src/components/islands/profile.jsx`)
* The dropdown indicator now always displays a visual cue: when closed,
a "hamburger" menu icon is shown, and when open, a "close" icon is
shown, improving clarity for users.
(`src/components/islands/profile.jsx`)
[[1]](diffhunk://#diff-6e8de8baae1c72ffb2bc7c8b736f1c8d029da6b9d6aeaf181d89dc09c1ef7498L56-R57)
[[2]](diffhunk://#diff-6e8de8baae1c72ffb2bc7c8b736f1c8d029da6b9d6aeaf181d89dc09c1ef7498L74-R90)

## Related Issue
<!-- Please prefix the issue number with Fixes/Resolves - example: Fixes
#123 or Resolves #123 -->

Follow-up to #247

## Screenshots/Screencasts
<!-- Please provide screenshots or video recording that demos your
changes (especially if it's a visual change) -->


[screen-capture.webm](https://github.com/user-attachments/assets/03d091d8-3f50-4522-aeda-379bde1b536e)

## Notes to Reviewer
<!-- Please state here if you added a new npm packages, or any extra
information that can help reviewer better review you changes -->

NA